+++ lib/src/analyzers/lint_analyzer/rules/rules_list/format_comment/format_comment_rule.dart
@@ -12,10 +12,9 @@ import '../../../models/severity.dart';
 import '../../models/common_rule.dart';
 import '../../rule_utils.dart';
 
+part 'config_parser.dart';
 part 'models/comment_info.dart';
-
 part 'models/comment_type.dart';
-
 part 'visitor.dart';
 
 class FormatCommentRule extends CommonRule {
@@ -23,8 +22,13 @@ class FormatCommentRule extends CommonRule {
 
   static const _warning = 'Prefer formatting comments like sentences.';
 
+  /// The patterns to ignore. They are used to ignore and not lint comments that
+  /// match at least one of them.
+  final Iterable<RegExp> _ignoredPatterns;
+
   FormatCommentRule([Map<String, Object> config = const {}])
-      : super(
+      : _ignoredPatterns = _ConfigParser.getIgnoredPatterns(config),
+        super(
           id: ruleId,
           severity: readSeverity(config, Severity.style),
           excludes: readExcludes(config),
@@ -32,7 +36,7 @@ class FormatCommentRule extends CommonRule {
 
   @override
   Iterable<Issue> check(InternalResolvedUnitResult source) {
-    final visitor = _Visitor()..checkComments(source.unit.root);
+    final visitor = _Visitor(_ignoredPatterns)..checkComments(source.unit.root);
 
     return [
       for (final comment in visitor.comments)