+++ lib/src/analyzers/lint_analyzer/rules/rules_list/format_comment/visitor.dart
@@ -11,6 +11,10 @@ const _ignoreExp = 'ignore:';
 const _ignoreForFileExp = 'ignore_for_file:';
 
 class _Visitor extends RecursiveAstVisitor<void> {
+  final Iterable<RegExp> _ignoredPatterns;
+
+  _Visitor(this._ignoredPatterns);
+
   final _comments = <_CommentInfo>[];
 
   Iterable<_CommentInfo> get comments => _comments;
@@ -54,8 +58,12 @@ class _Visitor extends RecursiveAstVisitor<void> {
 
     final isMacros = _regMacrosExp.hasMatch(text) || text == _macrosEndExp;
 
+    final isAnIgnoredPattern = _ignoredPatterns.any(
+      (regExp) => regExp.hasMatch(text),
+    );
+
     {
-      if (text.isEmpty || isIgnoreComment || isMacros) {
+      if (text.isEmpty || isIgnoreComment || isMacros || isAnIgnoredPattern) {
         return;
       } else {
         text = text.trim();